==> 0_LOADING.ipynb <==
## Loading (0)

import pandas as pd
import numpy as np 
import os
import json
import sqlite3
from datetime import datetime

---

### Function definitions

def path_files(path_directory):
    """
    Extracts all files names in a directory
    :param path_directory: path to the directory to extract files from
    :return: list of file paths in a given directory
    """
        
    directory = os.listdir(path_directory)
    json_files = []
    
    # Append all path files to list of json_files 

    for file in directory:
        if file.startswith('airlines-'):
            json_files.append(os.path.join(path_directory, file))

    return json_files

count = 0

def read_json(path_file: str):
    """
    Reads given json file, try to read entire file at once 
    :param path_files: path of the file
    :return: DataFrame of all tweets in given file
    """
    
    # Path counter
    
    global count
    print(str(count) + " " + path_file)
    count = count + 1
    
    # Try to read entire file at once, dtypes for tweet id's given to avoid JSON large number inaccuracy

    try:
        df_tweet_objects = pd.read_json(path_file, dtype={'id_str': object, 'in_reply_to_user_id_str': object, 
                                                          'in_reply_to_status_id_str': object, 'quoted_status_id_str': object})
    
    # When an error occurs, read file line by line 
    
    except ValueError:
        with open(path_file) as json_file:
            list_tweets = []
            for json_object in json_file:
                try:
                    tweet = json.loads(json_object)
                    list_tweets.append(tweet)
                except ValueError:
                    pass

            df_tweet_objects = pd.DataFrame(list_tweets)

    return df_tweet_objects.dropna(subset=['id_str', 'user'])

def tables(tweets):
    """
    Extracts nested dictionaries and
    :param tweets: DataFrame of all tweet information
    :return: one DataFrame for tweets and one DataFrame for users   
    """
    
    # Using pandas json_normalize to extract all keys from the nested user dictionary 
    
    users = pd.json_normalize(tweets['user'])
    
    # For loop to extract full_text from nested extended_tweet dictionary 

    for index, dictionary in enumerate(tweets['extended_tweet']):
        try:
            full_text = dictionary.get('full_text', np.nan)
            tweets.at[index, 'full_text'] = full_text
        except:
            pass
    
    # For loop to extract mentioned user id's from nested entities dictionary, return as tuple of user id's
    
    for index, dictionary in enumerate(tweets['entities']):
        mention_list = []
        for record in dictionary['user_mentions']:
            try:
                mentions = record.get('id_str', np.nan)
                mention_list.append(mentions)
            except:
                pass
        
        tweets.at[index, 'mentioned_users'] = ', '.join(filter(None, mention_list))
    
    tweets.rename(columns={'tweet_id': 'id_str'})

    # Match tweet and user id's to other tables, 'rename' columns and add empty columns as placeholders for later analysis
    
    tweets['tweet_id'] = tweets['id_str']        
    tweets['user_id'] = users['id_str']
    tweets['screen_name'] = users['screen_name']
    tweets['user_category'] = 'account'
    tweets['conversation_id'] = tweets['tweet_id']
    tweets['conversation_level'] = 0
    tweets['tokenized_text'] = 'unknown'
    tweets['created_at'] = pd.to_datetime(tweets['created_at'], infer_datetime_format=True)
    
    users['user_id'] = users['id_str']
    users['tweet_id'] = tweets['id_str']
    users['user_category'] = 'account'
    users['tweet_created_at'] = tweets['created_at']
    
    # Clean replies, quotes and tweet text by replacing nan values by proper values 
    
    tweets['reply_to_tweet'] = tweets['in_reply_to_status_id_str'].replace({np.nan: 0})
    tweets['reply_to_user'] = tweets['in_reply_to_user_id_str'].replace({np.nan: 0})
    tweets['quote_tweet'] = tweets['quoted_status_id_str'].replace({np.nan: 0})
    tweets['full_text'] = tweets['full_text'].fillna(tweets['text'])
    
    return tweets, users

def select(tweets, users):
    """
    Selects correct columns from DataFrames and cleans types
    :param tweets: DataFrame of all tweets
    :param users: DataFrame of all users
    :return: cleaned DataFrames 
    """
    
    # Copying useful columns into cleaned DataFrame 
    
    clean_tweets = tweets[['tweet_id', 'user_id', 'screen_name', 'user_category', 'full_text',
                           'tokenized_text', 'reply_to_tweet', 'reply_to_user', 'mentioned_users', 
                           'quote_tweet', 'conversation_id', 'conversation_level', 'lang', 'created_at']]
    
    clean_users = users[['user_id', 'tweet_id', 'name', 'screen_name', 'user_category', 'followers_count', 
                         'friends_count', 'statuses_count', 'verified', 'url', 'description', 'tweet_created_at']]
    
    # Defining correct types for columns in DataFrame 
    
    tweet_dict = {'tweet_id': int, 'user_id': int, 'screen_name': str, 'user_category': str, 'full_text': str, 'tokenized_text': str, 
                  'reply_to_tweet': int, 'reply_to_user': int, 'quote_tweet': int, 'mentioned_users': str, 
                  'conversation_id': str, 'conversation_level': int, 'lang': str, 'created_at': int}
    
    user_dict = {'user_id': int, 'tweet_id': int, 'name': str, 'screen_name': str, 'user_category': str,
                 'followers_count': int, 'friends_count': int, 'statuses_count': int, 'verified': bool, 
                 'url': str, 'description': str, 'tweet_created_at': int}
    
    clean_tweets = clean_tweets.astype(tweet_dict)
    clean_users = clean_users.astype(user_dict)
    
    # Adjusting created at columns to be in seconds 
    
    clean_tweets['created_at'] = clean_tweets['created_at'] // 1000000000
    clean_users['tweet_created_at'] = clean_users['tweet_created_at'] // 1000000000

    return clean_tweets, clean_users

def calculate_runtime(start, stop):
    """
    Calculates running time of every process. This cell is repeated in every notebook. 
    :param start: start time
    :param stop: stop time
    :return: output corresponding to running time for process 
    """
    
    runtime = stop - start
    
    days, seconds = runtime.days, runtime.seconds
    
    hours = days * 24 + seconds // 3600
    minutes = (seconds % 3600) // 60
    seconds = (seconds % 60)
    
    return 'Runtime loading {} files was: {} h, {} m {} s'.format(count, hours, minutes, seconds)

---

### Loading json files into database

Using seperate folders allows for easier monitoring of the loading process

conn_tweets = sqlite3.connect('tweets_airlines.db')

---

start = datetime.now()

---

All files are loaded in batches to reduce running time. From batch, a dataframe is created. This dateframe is then normalized, all useless columns are deleted and every column is given the correct type. 

If it is preferred to run all files at ones, only run the cells corresponding to data folder 1. 

### Data Folder 1 

files = path_files('./data/')

df_objects = pd.concat([read_json(i) for i in files], sort=False, ignore_index=True)

df_tweets, df_users = tables(df_objects)

df_tweets_clean, df_users_clean = select(df_tweets, df_users)

df_tweets_clean.to_sql('TWEETS', conn_tweets, if_exists="replace", index=False)

df_users_clean.to_sql('USERS', conn_tweets, if_exists="replace", index=False)

---

### Data Folder 2

files = path_files('./data_2/')

df_objects = pd.concat([read_json(i) for i in files], sort=False, ignore_index=True)

df_tweets, df_users = tables(df_objects)

df_tweets_clean, df_users_clean = select(df_tweets, df_users)

df_tweets_clean.to_sql('TWEETS', conn_tweets, if_exists="append", index=False)

df_users_clean.to_sql('USERS', conn_tweets, if_exists="append", index=False)

---

### Data Folder 3

files = path_files('./data_3/')

df_objects = pd.concat([read_json(i) for i in files], sort=False, ignore_index=True)

df_tweets, df_users = tables(df_objects)

df_tweets_clean, df_users_clean = select(df_tweets, df_users)

df_tweets_clean.to_sql('TWEETS', conn_tweets, if_exists="append", index=False)

df_users_clean.to_sql('USERS', conn_tweets, if_exists="append", index=False)

---

### Data Folder 4

files = path_files('./data_4/')

df_objects = pd.concat([read_json(i) for i in files], sort=False, ignore_index=True)

df_tweets, df_users = tables(df_objects)

df_tweets_clean, df_users_clean = select(df_tweets, df_users)

df_tweets_clean.to_sql('TWEETS', conn_tweets, if_exists="append", index=False)

df_users_clean.to_sql('USERS', conn_tweets, if_exists="append", index=False)

---

### Data Folder 5

files = path_files('./data_5/')

df_objects = pd.concat([read_json(i) for i in files], sort=False, ignore_index=True)

df_tweets, df_users = tables(df_objects)

df_tweets_clean, df_users_clean = select(df_tweets, df_users)

df_tweets_clean.to_sql('TWEETS', conn_tweets, if_exists="append", index=False)

df_users_clean.to_sql('USERS', conn_tweets, if_exists="append", index=False)

---

### Data Folder 6

files = path_files('./data_6/')

df_objects = pd.concat([read_json(i) for i in files], sort=False, ignore_index=True)

df_tweets, df_users = tables(df_objects)

df_tweets_clean, df_users_clean = select(df_tweets, df_users)

df_tweets_clean.to_sql('TWEETS', conn_tweets, if_exists="append", index=False)

df_users_clean.to_sql('USERS', conn_tweets, if_exists="append", index=False)

---

### Data Folder 7

files = path_files('./data_7/')

df_objects = pd.concat([read_json(i) for i in files], sort=False, ignore_index=True)

df_tweets, df_users = tables(df_objects)

df_tweets_clean, df_users_clean = select(df_tweets, df_users)

df_tweets_clean.to_sql('TWEETS', conn_tweets, if_exists="append", index=False)

df_users_clean.to_sql('USERS', conn_tweets, if_exists="append", index=False)

---

### Data Folder 8

files = path_files('./data_8/')

df_objects = pd.concat([read_json(i) for i in files], sort=False, ignore_index=True)

df_tweets, df_users = tables(df_objects)

df_tweets_clean, df_users_clean = select(df_tweets, df_users)

df_tweets_clean.to_sql('TWEETS', conn_tweets, if_exists="append", index=False)

df_users_clean.to_sql('USERS', conn_tweets, if_exists="append", index=False)

---

### Data Folder 9

files = path_files('./data_9/')

df_objects = pd.concat([read_json(i) for i in files], sort=False, ignore_index=True)

df_tweets, df_users = tables(df_objects)

df_tweets_clean, df_users_clean = select(df_tweets, df_users)

df_tweets_clean.to_sql('TWEETS', conn_tweets, if_exists="append", index=False)

df_users_clean.to_sql('USERS', conn_tweets, if_exists="append", index=False)

---

### Data Folder 10

files = path_files('./data_10/')

df_objects = pd.concat([read_json(i) for i in files], sort=False, ignore_index=True)

df_tweets, df_users = tables(df_objects)

df_tweets_clean, df_users_clean = select(df_tweets, df_users)

df_tweets_clean.to_sql('TWEETS', conn_tweets, if_exists="append", index=False)

df_users_clean.to_sql('USERS', conn_tweets, if_exists="append", index=False)

---

conn_tweets.close()


---


stop = datetime.now()

calculate_runtime(start, stop)
