==> 2_CONVERSATIONS.ipynb <==
## Conversations (2)

import sqlite3
import pandas as pd
import sys 
from datetime import datetime

pd.set_option('display.max_colwidth', None)  

---

conn_tweets = sqlite3.connect('tweets_airlines.db')

---

def calculate_runtime(start, stop):

    runtime = stop - start
    
    days, seconds = runtime.days, runtime.seconds
    
    hours = days * 24 + seconds // 3600
    minutes = (seconds % 3600) // 60
    seconds = (seconds % 60)
    
    return 'Runtime of creating Conversation Structure was: {} h, {} m {} s'.format(hours, minutes, seconds)

start = datetime.now()

---

### Conversation Structure

query_conv = '''SELECT tweet_id, user_id, reply_to_tweet, created_at 
                FROM TWEETS
                LIMIT 10000
                '''
df_conversations = pd.read_sql(query_conv, conn_tweets)

query_tweets = '''SELECT *
                  FROM TWEETS
                  LIMIT 10000
                  '''
df_tweets = pd.read_sql(query_tweets, conn_tweets)

query_users = '''SELECT *
                 FROM USERS
                 LIMIT 10000'''

df_users = pd.read_sql(query_users, conn_tweets)

        df_conversations.at[index, 'conversation_level'] = df_conversations.iloc[lambda x: x.index == row['reply_to_tweet'], 2].values[0] + 1


---

#### Giving all non roots the correct level and conversation id

Conversation level at start is 0 for all. For every non root, the conversation level is then defined as the conversation level of the tweet it is a reply to, + 1. 
Conversation ID is the tweet_id of the root of the conversation

df_conversations = df_conversations.sort_values(by='created_at')
df_conversations['conversation_level'] = 1
df_conversations['conversation_id'] = df_conversations['tweet_id'].astype(str)

df_tweets.shape[0]

roots = df_conversations['reply_to_tweet'] == 0
df_conversations.loc[roots, 'conversation_level'] = 0 

count = 0

for index, row in df_conversations[~roots].iterrows():
    
    count = count + 1
    
    parent_tweets = df_conversations.loc[df_conversations['tweet_id'] == row['reply_to_tweet']]
    
    if parent_tweets.size > 0:
        row_id = parent_tweets.iloc[0]['conversation_id']
        row_level = parent_tweets.iloc[0]['conversation_level'] + 1
        df_conversations.at[index,'conversation_level'] = row_level
        df_conversations.at[index,'conversation_id'] = row_id
    
    progress = '{count} / {total}'.format(count = count, total = (df_conversations[~roots].shape[0]))
    sys.stdout.write('\r' + progress)

df_conversations.shape[0]

---

#### Defining all conversation ID's correctly for level 0s and 1s

For root tweets, all conversation ID's are their tweet id with added .0. 

For all first replies, a dictionary counts the amount of level ones appear under one conversation ID. To the conversation_id of the level ones the string .X will be added, where X is a counter of the amount of 1s that still appear for the conversation_id that don't have a correctly defined conversation_id yet. The branches dictionary gets updated everytime a level 1 conversation_id is correctly defined. 

branches_dict = df_conversations[df_conversations['conversation_level'] == 1].groupby('conversation_id')['conversation_level'].count().to_dict()

def conv_id(conversation_id, level):
    
    if level == 0: 
        return conversation_id + '.0'

    else:
        if level == 1:
            try:
                level_ones = branches_dict.get(conversation_id)

                update_levels = level_ones - 1

                if update_levels == 0:
                    branches_dict.pop(conversation_id)

                else: 
                    branches_dict[conversation_id] = update_levels

                if level_ones > 1: 
                    return conversation_id + '.' + str(level_ones)

                else:
                    return conversation_id + '.1'
            except:
                pass
        else:
            return 'level > 1'

df_conversations['conversation_id'] = df_conversations.apply(lambda tweet: conv_id(tweet.conversation_id, tweet.conversation_level), axis=1)

---

#### Correcting conversation ID of tweets with level > 1 

For all tweets above level one the conversation ID is copied from the tweet it is a reply to. 

count = 0

for index, row in df_conversations[df_conversations['conversation_level'] > 1].iterrows():
    
    count = count + 1
    
    parent_tweets = df_conversations.loc[df_conversations['tweet_id'] == row['reply_to_tweet']]
    
    if parent_tweets.size > 1:
        row_id = parent_tweets.iloc[0]['conversation_id']
        df_conversations.at[index,'conversation_id'] = row_id
        
    progress = '{count} / {total}'.format(count = count, total = (df_conversations[df_conversations['conversation_level'] > 1].shape[0]))
    sys.stdout.write('\r' + progress)

---

dict_levels = df_conversations.set_index('tweet_id')['conversation_level'].to_dict()

dict_id = df_conversations.set_index('tweet_id')['conversation_id'].to_dict()

df_tweets['conversation_level'] = df_tweets['tweet_id'].map(dict_levels)

df_tweets['conversation_id'] = df_tweets['tweet_id'].map(dict_id)

df_tweets.to_sql('TWEETS', conn_tweets, if_exists="replace", index=False)

conn_tweets.close()

---

stop = datetime.now()

calculate_runtime(start, stop)

---

conn_tweets = sqlite3.connect('tweets_airlines.db')

---

### Conversation Structure

query_conv = '''SELECT tweet_id, reply_to_tweet, created_at 
                FROM TWEETS
                '''
df_conversations = pd.read_sql(query_conv, conn_tweets)

query_tweets = '''SELECT *
                  FROM TWEETS
                  '''
df_tweets = pd.read_sql(query_tweets, conn_tweets)

query_users = '''SELECT *
                 FROM USERS
                 '''
df_users = pd.read_sql(query_users, conn_tweets)

---

#### Giving all non roots the correct level and conversation id

Conversation level at start is 0 for all. For every non root, the conversation level is then defined as the conversation level of the tweet it is a reply to, + 1. 
Conversation ID is the tweet_id of the root of the conversation

---

df_conversations = df_conversations.sort_values(by='created_at')
df_conversations['conversation_level'] = 1
df_conversations['conversation_id'] = df_conversations['tweet_id'].astype(str)

roots = df_conversations['reply_to_tweet'] == 0
df_conversations.loc[roots, 'conversation_level'] = 0 

df_conversations = df_conversations.set_index('tweet_id')

roots = df_conversations['reply_to_tweet'] == 0
df_conversations.loc[roots, 'conversation_level'] = 0 

---

Splitting dataframes to reduce the runtime

df_conversations_1 = df_conversations.head(2000000).copy()
df_conversations_2_1 = df_conversations.head(4000000).copy()
df_conversations_2 = df_conversations_2_1.tail(2000000).copy()
df_conversations_3 = df_conversations.tail(2094135).copy()

count = 0

for index, row in df_conversations_1.iterrows():
    
    count = count + 1
    try:
        df_conversations_1.at[index, 'conversation_level'] = df_conversations_1.iloc[lambda x: x.index == row['reply_to_tweet'], 2].values[0] + 1
        df_conversations_1.at[index, 'conversation_id'] = df_conversations_1.iloc[lambda x: x.index == row['reply_to_tweet'], 3].values[0]
    except:
        pass
    
    progress = '{count} / {total}'.format(count = count, total = (df_conversations_1.shape[0]))
    sys.stdout.write('\r' + progress)

count = 0

for index, row in df_conversations_2.iterrows():
    
    count = count + 1
    try:
        df_conversations_2.at[index, 'conversation_level'] = df_conversations_2.iloc[lambda x: x.index == row['reply_to_tweet'], 2].values[0] + 1
        df_conversations_2.at[index, 'conversation_id'] = df_conversations_2.iloc[lambda x: x.index == row['reply_to_tweet'], 3].values[0]
    except:
        pass
    
    progress = '{count} / {total}'.format(count = count, total = (df_conversations_2.shape[0]))
    sys.stdout.write('\r' + progress)

count = 0

for index, row in df_conversations_3.iterrows():
    
    count = count + 1
    try:
        df_conversations_3.at[index, 'conversation_level'] = df_conversations_3.iloc[lambda x: x.index == row['reply_to_tweet'], 2].values[0] + 1
        df_conversations_2.at[index, 'conversation_id'] = df_conversations_2.iloc[lambda x: x.index == row['reply_to_tweet'], 3].values[0]        
    except:
        pass
    
    progress = '{count} / {total}'.format(count = count, total = (df_conversations_3.shape[0]))
    sys.stdout.write('\r' + progress)

df_conversations = pd.concat([df_conversations_1, df_conversations_2, df_conversations_3])

---

#### Defining all conversation ID's correctly for level 0s and 1s

For root tweets, all conversation ID's are their tweet id with added .0. 

For all first replies, a dictionary counts the amount of level ones appear under one conversation ID. To the conversation_id of the level ones the string .X will be added, where X is a counter of the amount of 1s that still appear for the conversation_id that don't have a correctly defined conversation_id yet. The branches dictionary gets updated everytime a level 1 conversation_id is correctly defined. 

branches_dict = df_conversations[df_conversations['conversation_level'] == 1].groupby('conversation_id')['conversation_level'].count().to_dict()

def conv_id(conversation_id, level):
    """
    Updates conversation_id for every record. Takes the amount of level 1's per conversation_id and creates a unique code for every of them
    :param conversation_id: current conversation_id
    :param level: conversation_level 
    :return: updated conversation_id 
    """
    
    # All level 0 tweets' conversation_id get the addition .0
    if level == 0: 
        return conversation_id + '.0'

    else:
        # All level 1 tweets' conversation_id get a unique addition
        if level == 1:
            try:
                level_ones = branches_dict.get(conversation_id)

                update_levels = level_ones - 1

                if update_levels == 0:
                    branches_dict.pop(conversation_id)

                else: 
                    branches_dict[conversation_id] = update_levels

                if level_ones > 1: 
                    return conversation_id + '.' + str(level_ones)

                else:
                    return conversation_id + '.1'
            except:
                pass
        # All tweets above level 1 get no conversation_id, this is corrected later 
        else:
            return 'level > 1'

df_conversations['conversation_id'] = df_conversations.apply(lambda tweet: conv_id(tweet.conversation_id, tweet.conversation_level), axis=1)

---

#### Correcting conversation ID of tweets with level > 1 

For all tweets above level one the conversation ID is copied from the tweet it is a reply to. 

df_conversations_1 = df_conversations.head(2000000).copy()
df_conversations_2_1 = df_conversations.head(4000000).copy()
df_conversations_2 = df_conversations_2_1.tail(2000000).copy()
df_conversations_3 = df_conversations.tail(2094135).copy()

count = 0

for index, row in df_conversations_1[df_conversations_1['conversation_level'] > 1].iterrows():
    
    count = count + 1
    
    try:
        df_conversations_1.at[index, 'conversation_id'] = df_conversations_1.iloc[lambda x: x.index == row['reply_to_tweet'], 3].values[0]
    except:
        pass
    
    progress = '{count} / {total}'.format(count = count, total = (df_conversations_1[df_conversations_1['conversation_level'] > 1].shape[0]))
    sys.stdout.write('\r' + progress)

count = 0

for index, row in df_conversations_2[df_conversations_2['conversation_level'] > 1].iterrows():
    
    count = count + 1
    
    try:
        df_conversations_2.at[index, 'conversation_id'] = df_conversations_2.iloc[lambda x: x.index == row['reply_to_tweet'], 3].values[0]
    except:
        pass    
        
    progress = '{count} / {total}'.format(count = count, total = (df_conversations_2[df_conversations_2['conversation_level'] > 1].shape[0]))
    sys.stdout.write('\r' + progress)

count = 0

for index, row in df_conversations_3[df_conversations_3['conversation_level'] > 1].iterrows():
    
    count = count + 1
    
    try:
        df_conversations_3.at[index, 'conversation_id'] = df_conversations_3.iloc[lambda x: x.index == row['reply_to_tweet'], 3].values[0]
    except:
        pass    
        
    progress = '{count} / {total}'.format(count = count, total = (df_conversations_3[df_conversations_3['conversation_level'] > 1].shape[0]))
    sys.stdout.write('\r' + progress)

df_conversations = pd.concat([df_conversations_1, df_conversations_2, df_conversations_3])

---

dict_levels = df_conversations['conversation_level'].to_dict()

dict_id = df_conversations['conversation_id'].to_dict()

df_tweets['conversation_level'] = df_tweets['tweet_id'].map(dict_levels)

df_tweets['conversation_id'] = df_tweets['tweet_id'].map(dict_id)

df_tweets.to_sql('TWEETS', conn_tweets, if_exists="replace", index=False)

conn_tweets.close()

---

stop = datetime.now()

calculate_runtime(start, stop)
